requirements.txt
Django==1.11.1
factory-boy==2.8.1
Faker==0.7.12
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
